Hooks for clean code in the page of trader registration and smooth scroll when action occurs in form
