+++ src/components/forms/trade-buyer/NavigationButtons.tsx
@@ -15,13 +15,34 @@ const NavigationButtons: React.FC<NavigationButtonsProps> = ({
   showPrev = true,
   showNext = true,
 }) => {
+
+  const scrollToTop = () => {
+    window.scrollTo({
+      top: 0,
+      behavior: 'smooth',
+    });
+  };
+
+  // Wrapping functions to add scroll functionality
+  const handlePrevClick = () => {
+    if (handlePrev) {
+      handlePrev();
+    }
+    scrollToTop();
+  };
+
+  const handleNextClick = () => {
+    handleNext();
+    scrollToTop();
+  };
+
   return (
     <nav className="flex flex-wrap gap-10 justify-between items-center self-end mt-11 max-w-full text-lg font-bold leading-loose text-white whitespace-nowrap w-[554px] max-md:mt-10">
       {showPrev && handlePrev && (
         <button
           className="flex flex-col self-stretch my-auto w-[53px] h-[53px] bg-stone-600"
           aria-label="Previous"
-          onClick={handlePrev}
+          onClick={handlePrevClick}  // Use wrapped function
         >
           <span className="self-center my-auto">&lt;</span>
         </button>
@@ -29,7 +50,7 @@ const NavigationButtons: React.FC<NavigationButtonsProps> = ({
       {showNext && (
         <button
           className="flex flex-col self-stretch my-auto w-[171px] px-12 py-3.5 bg-stone-600 min-h-[53px] max-md:px-5"
-          onClick={handleNext}
+          onClick={handleNextClick}  // Use wrapped function
         >
           {isSubmit ? 'SUBMIT' : 'NEXT'}
         </button>
